Use algo in adapter
