+++ crates/fuel-core/benches/gas_price_algo.rs
@@ -3,7 +3,7 @@ use criterion::{
     criterion_main,
     Criterion,
 };
-// TODO: Move from `gas_price_testground`
+// TODO: Move from `gas-price-algorithm`
 fn gas_price_algo(_c: &mut Criterion) {}
 
 criterion_group!(benches, gas_price_algo);
