+++ crates/gas-price-algorithm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "gas_price_testground"
+name = "gas-price-algorithm"
 authors.workspace = true
 categories.workspace = true
 edition.workspace = true
@@ -8,6 +8,15 @@ keywords.workspace = true
 license.workspace = true
 repository.workspace = true
 version.workspace = true
+
+
+[lib]
+name = "gas_price_algorithm"
+path = "src/lib.rs"
+
+[[bin]]
+name = "main"
+path = "src/main.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
